Fixing dependencies for moving FeedViewState.
