+++ android/app/build.gradle.kts
@@ -120,8 +120,10 @@ dependencies {
     implementation(project(":android:design-system"))
     implementation(project(":android:widgets"))
     implementation(project(":core:displaymodels"))
+    implementation(project(":core:feature"))
     implementation(project(":core:models"))
     implementation(project(":data:local-sqldelight"))
+    implementation(project(":feature:feed"))
     implementation(project(":shared"))
     implementation(libs.accompanist.flowlayout)
     implementation(libs.accompanist.placeholder.material)
